IPL Events at Ace of Pubs
Get ready to put your IPL knowledge to the test with Ace of Pubs' IPL Quiz events! Join fellow cricket enthusiasts for a night of fun, interactive, and challenging trivia. With a variety of topics covered and free drinks up for grabs, this is an event you don't want to miss. Sign up now and show off your skills!